requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy==2.0.40
pydantic
python-multipart
# Flask dependencies (can be removed later if Flask is fully replaced)
# Flask==3.1.0
# Flask-SQLAlchemy==3.1.1
# Werkzeug==3.1.3
# Jinja2==3.1.6
# Flask-Cors==4.0.0
# Dependencies potentially needed for models/utils
PyMySQL==1.1.1
cryptography==36.0.2
PyJWT==2.8.0

annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==36.0.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mercadopago==2.3.0
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.0.1
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==36.0.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mercadopago==2.3.0
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.0.1
passlib[bcrypt]
python-jose[cryptography]
authlib
google-auth-oauthlib
